node2nix packages: Replace tgz reference with symbolic one.
This is almost equivalent except that with symbolic reference node2nix
consults the NPM registry for package.json information instead of the
package.json in the tgz file. The registry canonicalizes the
"dependencies" attribute in that it adds all "optionalDependencies",
which the npm binary would do for local package.json files, but
node2nix does not.

For spark-wallet optionalDependencies are not listed in the tgz
distributed package.json, hence node2nix misses them. This leads to a
missing qrcode-terminal package and spark-wallet dying when called
with -Q.